backend fix

+++ backend/src/main/java/com/example/demo/interfaces/ICategoriaProductoService.java
@@ -14,5 +14,5 @@ public interface ICategoriaProductoService {
 
 	void deleteById(long id);
 
-	List<ProductoEntity> findProductoByCategoriaId(int categoriaId);
+	List<ProductoEntity> findProductoByCategoriaId(Long categoriaId);
 }
